added connectionMessages file
